feat(web): update MigrationConfig for cross-platform support (#42)
- Decompose 1017-line MigrationConfig.tsx into subcomponents:
  - StrategySection: migration strategy selection with platform-conditional descriptions
  - ObjectSelectionSection: table/object selection with filters
  - TargetTierSection: target service tier with PG services (Burstable, General Purpose, Memory Optimized)
  - RegionPricingSection: region and pricing display
  - ScriptSection: pre/post migration scripts
  - ReviewApproveSection: review and approve/save plan
  - index.tsx: main orchestrator composing all subcomponents
- Add PostgreSQL Flexible Server tiers alongside existing SQL Server services
- Platform-conditional strategy descriptions (Change Tracking vs Logical Replication)
- Pass sourcePlatform/targetPlatform in plan save/update API calls
- Platform-aware target database labels and server placeholders
